split testprojection code
